requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==4.2.16
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
gunicorn==20.1.0
idna==3.10
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
